修复了一处bug。内存模块应该先于 线程模块。debug了好久，发现 mutex_lock 时无法P操作，因为内存模块还未初始化锁。

+++ src/lib/math.h
@@ -1,7 +1,7 @@
 #ifndef __LIB_MATH_H
 #define __LIB_MATH_H
 
-int ceil(double x);
+double ceil(double x);
 
 int pow(int x, int y);
